app/requirements.txt
numpy
streamlit
opencv-python
Pillow
keras==3.3.3
tensorflow==2.16.1